requirements.txt
numpy
pillow
cryptography